Relax Default impls
Deriving them adds a bound of `K: Default` (and `V`, `T` respectively)
as it won't inspect the field types precisely to note that they have
more generic impls, too. Also the macro won't make the exact field types
part of the public API.

By manually implementing them it is possible to default-construct sets
and maps of non-Default types with similar semantics as vector: as an
empty container.